+++ CHANGELOG.md
@@ -17,6 +17,7 @@ CHANGELOG
  * Add `framework.html_sanitizer.sanitizers.*.default_action` config option
  * Deprecate parameters `router.request_context.scheme` and `router.request_context.host`;
    use the `router.request_context.base_url` parameter or the `framework.router.default_uri` config option instead
+ * Add decoration stack to `debug:container` command output
 
 8.0
 ---